requirements.txt
Flask
torch>=1.7.0
torchvision>=0.8.0
numpy
opencv-python
Pillow
matplotlib
scikit-learn
requests
# Add any other dependencies your app.py might need
